# springdoc-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/status/standard/create": {
      "post": {
        "tags": [
          "Status Controller"
        ],
        "summary": "Create a standard status",
        "description": "Creates a new standard status record in the system with the provided details.",
        "operationId": "createStandardStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStandardStatusRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperStandardStatusResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/standard/{standardId}/sections/{sectionId}/upload": {
      "post": {
        "tags": [
          "Standards"
        ],
        "summary": "Upload a file to a section",
        "description": "Uploads a single PDF file to the specified section. Multiple files per request are not allowed.",
        "operationId": "uploadFileToSection",
        "parameters": [
          {
            "name": "standardId",
            "in": "path",
            "description": "ID of the standard",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sectionId",
            "in": "path",
            "description": "ID of the section",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "User ID of the uploader",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "description": "File to upload (PDF only)",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperFileStorageResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/standard/{standardId}/sections/create": {
      "post": {
        "tags": [
          "Standards"
        ],
        "summary": "Create a section under a standard",
        "description": "Adds a new section to a specific ISO standard identified by its ID.",
        "operationId": "createSection",
        "parameters": [
          {
            "name": "standardId",
            "in": "path",
            "description": "ID of the parent standard",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStandardSectionRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperStandardSectionResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/standard/create": {
      "post": {
        "tags": [
          "Standards"
        ],
        "summary": "Create a new standard",
        "description": "Creates a new ISO standard using the provided details.",
        "operationId": "createStandard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStandardRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperStandardResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/standard/update": {
      "patch": {
        "tags": [
          "Standards"
        ],
        "summary": "Update a standard",
        "description": "Update an ISO standard using the provided details.",
        "operationId": "updateStandard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStandardRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperStandardResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/standard/{standardId}/treeView": {
      "get": {
        "tags": [
          "Standards"
        ],
        "summary": "Get standard section tree",
        "description": "Fetches the hierarchical tree of all sections for a given standard.",
        "operationId": "getStandardSectionTree",
        "parameters": [
          {
            "name": "standardId",
            "in": "path",
            "description": "ID of the standard",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperStandardSectionTreeDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/standard/all": {
      "get": {
        "tags": [
          "Standards"
        ],
        "summary": "Get all standards with pagination",
        "description": "Retrieves a paginated list of all ISO standards. Default page=0, size=20.",
        "operationId": "getAllStandards",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperListStandardResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/standard/{standardId}/sections/{sectionId}/files/{fileId}": {
      "delete": {
        "tags": [
          "Standards"
        ],
        "summary": "Delete a file from a section",
        "description": "Removes a specific file by ID from the given section.",
        "operationId": "deleteFileFromSection",
        "parameters": [
          {
            "name": "standardId",
            "in": "path",
            "description": "Standard ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sectionId",
            "in": "path",
            "description": "Section ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "fileId",
            "in": "path",
            "description": "File ID to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperString"
                }
              }
            }
          }
        }
      }
    },
    "/api/standard/delete/{standardId}": {
      "delete": {
        "tags": [
          "Standards"
        ],
        "summary": "Delete a standard",
        "description": "Deletes an existing ISO standard by its ID.",
        "operationId": "deleteStandard",
        "parameters": [
          {
            "name": "standardId",
            "in": "path",
            "description": "ID of the standard to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperString"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateStandardStatusRequestDTO": {
        "required": [
          "statusName"
        ],
        "type": "object",
        "properties": {
          "statusName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ApiResponseWrapperStandardStatusResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/StandardStatusResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StandardStatusResponseDTO": {
        "type": "object",
        "properties": {
          "statusId": {
            "type": "integer",
            "format": "int32"
          },
          "statusName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseWrapperFileStorageResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/FileStorageResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FileStorageResponseDTO": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "integer",
            "format": "int32"
          },
          "fileName": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "uploadedByUserId": {
            "type": "integer",
            "format": "int32"
          },
          "uploadedByCustomerId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CreateStandardSectionRequestDTO": {
        "required": [
          "number",
          "title"
        ],
        "type": "object",
        "properties": {
          "standardId": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "parentSectionId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseWrapperStandardSectionResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/StandardSectionResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StandardSectionResponseDTO": {
        "type": "object",
        "properties": {
          "sectionId": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "standardId": {
            "type": "integer",
            "format": "int32"
          },
          "parentSectionId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CreateStandardRequestDTO": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "publishedDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "ApiResponseWrapperStandardResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/StandardResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StandardResponseDTO": {
        "type": "object",
        "properties": {
          "standardId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "publishedDate": {
            "type": "string",
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseWrapperString": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PatchFieldInteger": {
        "type": "object",
        "properties": {
          "provided": {
            "type": "boolean"
          }
        }
      },
      "PatchFieldString": {
        "type": "object",
        "properties": {
          "provided": {
            "type": "boolean"
          }
        }
      },
...
      "UpdateStandardRequestDTO": {
        "type": "object",
        "properties": {
          "standardId": {
            "$ref": "#/components/schemas/PatchFieldInteger"
          },
          "name": {
            "$ref": "#/components/schemas/PatchFieldString"
          },
          "version": {
            "$ref": "#/components/schemas/PatchFieldString"
          }
        }
      },
...
      "ApiResponseWrapperStandardSectionTreeDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/StandardSectionTreeDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SectionNodeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "StandardSectionTreeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionNodeDTO"
            }
          },
          "number": {
            "type": "string"
          }
        }
      },
      "ApiResponseWrapperListStandardResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StandardResponseDTO"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}